django_backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
openpyxl==3.1.5
pytz==2025.2